+++ apps/cms-v2/src/modules/datocms/datocms-bulk-sync-processor.ts
@@ -1,0 +1,50 @@
+import { BulkImportProductFragment } from "../../../generated/graphql";
+import { BulkSyncProcessor, BulkSyncProcessorHooks } from "../bulk-sync/bulk-sync-processor";
+import { DatocmsProviderConfigType } from "../configuration";
+import { DatoCMSClientBrowser } from "./datocms-client-browser";
+
+export class DatocmsBulkSyncProcessor implements BulkSyncProcessor {
+  constructor(private config: DatocmsProviderConfigType) {}
+
+  async uploadProducts(
+    products: BulkImportProductFragment[],
+    hooks: BulkSyncProcessorHooks
+  ): Promise<void> {
+    const client = new DatoCMSClientBrowser({
+      apiToken: this.config.authToken,
+    });
+
+    products.flatMap((product) =>
+      product.variants?.map((variant) => {
+        if (hooks.onUploadStart) {
+          hooks.onUploadStart({ variantId: variant.id });
+        }
+
+        return client
+          .upsertProduct({
+            configuration: this.config,
+            variant: {
+              id: variant.id,
+              name: variant.name,
+              channelListings: variant.channelListings,
+              product: {
+                id: product.id,
+                name: product.name,
+                slug: product.slug,
+              },
+            },
+          })
+          .then((r) => {
+            if (hooks.onUploadSuccess) {
+              hooks.onUploadSuccess({ variantId: variant.id });
+            }
+          })
+          .catch((e) => {
+            if (hooks.onUploadError) {
+              hooks.onUploadError({ variantId: variant.id, error: e });
+            }
+          });
+      })
+    );
+  }
+}
